bgp-orchestrator/backend/requirements.txt
# BGP Orchestrator Backend Requirements
# Production-grade dependencies with pinned versions

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis
redis==5.0.1

# Observability
prometheus-client==0.19.0

# Logging
structlog==23.2.0

# BGP & Network Automation
pybgpstream==2.2.0
pysros==24.7.1
ncclient==0.6.15

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# HTTP Client
httpx==0.25.1

# Utilities
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Testing
fakeredis==2.20.1

# Machine Learning & AI
xgboost==2.0.3
scikit-learn==1.3.2
prophet==1.1.5
pandas==2.1.4
numpy==1.26.2

# ONNX Runtime (for fast ML inference)
onnx==1.15.0
onnxruntime==1.16.3
skl2onnx==1.15.0

# Time Series Analysis
statsmodels==0.14.0

# Continuous Profiling
pyroscope-io==0.7.2

# Kafka Streaming
aiokafka==0.11.0

# Feature Store
feast[redis]==0.36.0
pyarrow==14.0.1